split the test router
